% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createHicPainteRObJ.R
\name{createHicPainteRObj}
\alias{createHicPainteRObj}
\title{Create an hicPainteR object}
\usage{
createHicPainteRObj((contact,mapType,Name,zoom=NA,high_fact=1e6,pCol="red",alpha=0.05,enhance=NA,colBin=500,
       use_ramp=FALSE,ramp=NA,log=FALSE,pchP=".",lwdP=2,ltyP=1,cexP=2,scores=NA,use.scores=FALSE,...))
}
\arguments{
\item{contact}{GenomicInteractions object}

\item{mapType}{character - possible choises are cMAP, TAD, Loops}

\item{Name}{character - is the name of the object ( default NA )}

\item{high_fact}{numeric - value for manage the proportion of graph ( default 1e6 ) (TODO control if making it automatic)}

\item{pCol}{charachter - color used if  use_ramp == FALSE ( default "red" )}

\item{enhance}{numeric - transform factor for counts on cMap ( default = NA )}

\item{colBin}{numeric - used to calculate the number of the palette when use_ramp == TRUE ( default  500 )}

\item{use_ramp}{logic - select to switch from flat color to colorRampPalette ( default FALSE )}

\item{ramp}{colorRampPalette object - palette used when use_ramp == TRUE, when is set to NA ramp= ( default NA )}

\item{log}{logic - transform the counts of the contact object in logaritmic ( default NA )}

\item{pchP}{character - pch code for points drawing}

\item{cexP}{numeric - cex value for point drawing}

\item{lwdP}{numeric - lwd thick for line drawing}

\item{ltyP}{numeric - lty code for line drawing}
}
\description{
create an hicPainteR object starting from a GenomicInteractions object
}
\examples{
 ...

}
\references{
...
}
\seealso{
...
}
