% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gInteractionsTocMap.R
\name{gInteractionsTocMap}
\alias{gInteractionsTocMap}
\title{Genomics Interaction to Genomic Map}
\usage{
gInteractionsTocMap( GI, zoom = NA )
}
\arguments{
\item{GI}{GenomicInteractions object}

\item{zoom}{character - genomic coordinate in wich the GRanges will be contruct, if not privide it will use all the contact of GI.
The format need to be "chromosome:start-end" ( default = NA )}
}
\description{
transform a GenomicsInteraction object to a Granges with mCol "edge" and "Ys" that will be use for create a hicPainter Object,
if zoom is provided it will extract only this portion of genome
}
\examples{
 ...

}
\references{
...
}
\seealso{
...
}
